[CBR-146] BREAKS THE BUILD: Refactor Inductive
This turns `Inductive` to a list of wallet events. We needed this in some
places already, and ended up with two variations on `Inductive`, which wasn't
very satisfactory. Moreover, the blocktree generator probably needs yet another
variation again. With this new definition they can all use the same construct.
It also happens to clear up a nasty bit of code in the DSL-versus-cardano
wallet interpreter.

This does not update all the code, since I'm planning to replace the
`Inductive` generator.